DispSync: Workaround HH hardware vsync issue
The HWC of HH seems to have a bug where it fails to generate vsync events
upon the first eventControl call to turn on HWC_EVENT_VSYNC.

The workaround is for DispSync to detect when it has received no resync
samples after many presents, and return false so that SF will disable HW
vsync and re-enable it in the next present. This gets the HWC of HH out of
the bad state.

Bug: 25845510
Signed-off-by: Haixia Shi <hshi@google.com>
Signed-off-by: Alex Saiko <solcmdr@gmail.com>